requirements.txt
fastapi>=0.95.1
uvicorn>=0.21.1
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.10
python-dotenv>=1.0.0
loguru>=0.7.0
openai>=1.0.0
duckduckgo-search>=4.1.1
